Fix failing downloads of volatile files
Tasks waiting for a download of a volatile file (e.g., InRelease,
Packages.gz, etc.) are now notified by the downloading task if the file
hasn't changed on the mirror. Previously, these tasks weren't notified
in this case, resulting in the download being aborted. Additionally, all
tasks with the status ActiveDownloadStatus::Init now wait 5 seconds for
the download to start before aborting it. This avoids unnecessary
download errors on slow connections.